requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-address==0.2.8
django-bootstrap-form==3.4
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
